Removed pin out number from wire message: the output pin number is always 0 because it comes out from a fundamental logic gate which always has one output. Renamed some nodes in composite components to indicate the input pin they correspond to. Added up counter class. Added new ComponentArray constructor which allows get as a name a vector of strings. This is particularly useful when the ComponentArray is made up of input nodes, for which naming is important to be individually set arbitrarily (instead of "name0", "name1", etc. as before)

+++ Digital circuit simulator/DCSEngine.hpp
@@ -30,7 +30,7 @@ public:
 	static void addComponent(DCSComponent* component);
 	static void addInput(DCSInput* input);
 	static void addWire(DCSWire* p_wire);
-	static void run(int steps=10);
+	static void run(uint64_t steps=10);
 	static void initialize(std::vector<DCSComponent*> cVec = inputVector);
 	static int getClockPeriod();
 	static int getStepNumber();
